Move helpers and fixtures into conftest.py
- modified:   test/bathtub_pkg_test_suite/pkg_test.py
- renamed:    test/bathtub_pkg_test_suite/pkg_test.py -> test/conftest.py
  - Refactored helpers and fixtures into `conftest.py` module.
  - New `svunit` fixture.
